TDDに基づくモジュール化リファクタリングの実施
## 概要
- アーキテクチャを大幅にリファクタリング
- 関心の分離に基づいて7つのモジュールに分割
- TDD（Red → Green → Refactor）に厳守して開発

## 主な変更

### ディレクトリ構造
```
src/
  ├── logger.js         - ログ出力機能
  ├── config.js         - 設定管理（YAML）
  ├── browser.js        - ブラウザ操作
  ├── repository.js     - GitHubリポジトリ取得
  ├── html.js           - HTML生成
  ├── port.js           - ポート検索
  └── server.js         - HTTPサーバー実装

tests/
  ├── logger.test.js
  ├── config.test.js
  ├── browser.test.js
  ├── repository.test.js
  ├── html.test.js
  ├── port.test.js
  └── server.test.js
```

### 実装詳細
1. **logger.js**: 時刻フォーマット、ログ出力の責務
2. **config.js**: YAML設定ファイルの読み込み・保存
3. **browser.js**: プラットフォーム別ブラウザ起動
4. **repository.js**: gh CLIを通じたGitHub連携
5. **html.js**: HTML生成（テンプレート）
6. **port.js**: 利用可能ポートの検索・確保
7. **server.js**: HTTPサーバー実装とエンドポイント管理

### テスト結果
- テストスイート: 7個（全てPASS）
- テストケース: 62個（全てPASS）
- カバレッジ: 65.38%（ステートメント）
- 実行時間: 0.392秒

### Jest設定
- testMatch: tests/**/*.test.js
- testEnvironment: node
- coverage対象: src/**/*.js

### パッケージ更新
- version: 1.0.0 → 2.0.0
- description: 詳細なプロジェクト説明に更新
- keywords: github, repository, observer を追加
- engines: Node.js 20.0.0以上を指定

## TDD プロセス
各モジュール実装時に以下のサイクルを厳守：
1. RED: テストケースを先に実装
2. GREEN: 最小限の実装でテストをPASS
3. REFACTOR: コード品質を向上

このアプローチにより、
- テストカバレッジが高い
- 実装品質が向上
- 保守性と拡張性が向上
