+++ jest.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  testEnvironment: 'node',
+  collectCoverageFrom: [
+    'src/**/*.js',
+    '!src/**/*.test.js',
+    '!src/**/index.js'
+  ],
+  coverageDirectory: 'coverage',
+  testMatch: [
+    '**/tests/**/*.test.js'
+  ]
+};
